still buggy but the basic functionality of collapse transitions is there
