+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectColumn.css
@@ -34,6 +34,8 @@
     padding-top: 8px;
     padding-right: 8px;
     box-sizing: border-box;
+    min-height: 48px; /* would be nice if this was auto */
+    transition: min-height 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms; 
 }
 
 .expanded-row {
